创建独立的domain层管理数据库
